# src/main/resources/swagger.yml
openapi: 3.0.0
info:
  title: API de Libros
  description: API para gestionar libros en el sistema.
  version: 1.0.0

servers:
  - url: http://localhost:8080
    description: Servidor de desarrollo

paths:
  /api/libros:
    get:
      summary: Obtener todos los libros
      operationId: getAllLibros
      tags:
        - Libros
      responses:
        '200':
          description: Un arreglo JSON de libros
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Libro'
    post:
      summary: Crear un nuevo libro
      operationId: createLibro
      tags:
        - Libros
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Libro'
      responses:
        '201':
          description: Libro creado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Libro'

  /api/libros/{id}:
    get:
      summary: Obtener un libro por ID
      operationId: getLibroById
      tags:
        - Libros
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Un objeto de libro
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Libro'
        '404':
          description: Libro no encontrado
    put:
      summary: Actualizar un libro por ID
      operationId: updateLibro
      tags:
        - Libros
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Libro'
      responses:
        '200':
          description: Libro actualizado exitosamente
        '404':
          description: Libro no encontrado
    delete:
      summary: Eliminar un libro por ID
      operationId: deleteLibro
      tags:
        - Libros
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Libro eliminado exitosamente
        '404':
          description: Libro no encontrado

components:
  schemas:
    Libro:
      type: object
      properties:
        id_libro:
          type: integer
          description: El ID del libro.
        titulo:
          type: string
          description: El título del libro.
        autor:
          type: string
          description: El autor del libro.
        precio:
          type: number
          format: float
          description: El precio del libro.
      required:
        - titulo
        - autor
        - precio